Extract reader and writer logic
DBReader and DBWriter logic moved
to its own files to improve
readability.
